added profiles table and functionality to friends list
added a table for profiles to the database. also added an item in friends list to show if the friend has a profile posted or not, and added the option to view a profile if it is posted. the actual printing of a profile has not been implemented yet though.